factor out import ops into its own trait and fix IndexedGraph to no longer implement deref so we don't break it accidentally
